requirements.txt
Django==2.0.3

pytz==2018.3
httplib2==0.10.3
feedparser==5.2.1
Markdown==2.6.11
psycopg2==2.7.4
Pillow==5.0.0
versiontools==1.9.1
statsd==3.2.2
pep8==1.7.1
pyflakes==1.6.0
mccabe==0.6.1
enum34==1.1.6
configparser==3.5.0
pycodestyle==2.3.1
flake8==3.5.0
tornado==4.5.3 # pyup: <5.0
beautifulsoup4==4.6.0
cssselect==1.0.3
lxml==4.1.1
fuzzywuzzy==0.16.0
sure==1.4.7
traitlets==4.3.2  # ipython
ipython_genutils==0.2.0  # ipython
decorator==4.2.1  # ipython
ptyprocess==0.5.2  # ipython
pexpect==4.4.0  # ipython
pathlib2==2.3.0  # ipython
pickleshare==0.7.4  # ipython
simplegeneric==0.8.1  # ipython
path.py==11.0  # ipython
backports.shutil_get_terminal_size==1.0.0  # ipython
wcwidth==0.1.7  # ipython
prompt_toolkit==1.0.15  # ipython
pygments==2.2.0  # ipython
ipython==5.5.0  # pyup: <6.0.0
ipdb==0.11
rdflib==4.2.2
selenium==3.11.0
coverage==4.5.1
pyasn1==0.4.2
pyOpenSSL==17.5.0
ndg-httpsclient==0.4.4
requests==2.18.4
logilab-common==1.4.1
logilab-astng==0.24.3
astroid==1.6.1
pylint==1.8.2
six==1.11.0
ipaddress==1.0.19  # faker
faker==0.8.12  # factory_boy
factory_boy==2.10.0
ldap3==2.4.1
sqlparse==0.2.4
contextlib2==0.5.5
boto==2.48.0
python-dateutil==2.7.0

djangowind==1.1.0
django-indexer==0.3.0
django-templatetag-sugar==1.0
django-paging==0.2.5
django-annoying==0.10.4
django-appconf==1.0.2
rcssmin==1.0.6  # django-compressor
rjsmin==1.0.12  # django-compressor
django-compressor==2.2
django-statsd-mozilla==0.4.0
raven==6.6.0
django-bootstrap4==0.0.6
django-debug-toolbar==1.9.1
django-waffle==0.14.0
django-jenkins==0.110.0
django-smoketest==1.1.0
django-extensions==2.0.5
django-stagingcontext==0.1.0
django-ga-context==0.1.0
django-impersonate==1.3
django-markwhat==1.6.0
gunicorn==19.7.1
django-infranil==1.1.0
django-flatblocks==0.9.4
django-storages-redux==1.3.3
django-treebeard==4.2.1
django-modelcluster==4.1

wagtail==2.0
djangorestframework==3.7.7
Willow==1.1
django-modelcluster==4.1
django-taggit==0.22.2
Unidecode==1.0.22
html5lib==1.0.1
typing==3.6.4 # needed for django-extensions
webencodings==0.5.1 # needed for wagtail
urllib3==1.22 # wagtail elasticsearch backend

django-s3sign==0.1.4
django-cacheds3storage==0.1.2
ccnmtlsettings==1.3.0
text_unidecode==1.2

cryptography==2.1.4

pbr==3.1.1
PyYAML>=3.10.0 # MIT
stevedore>=1.20.0 # Apache-2.0
bandit==1.4.0

